Disabled std::is_trivially_copyable == llvm::is_trivially_copyable consistency assert.
This assert broke in 16.7 https://developercommunity.visualstudio.com/content/problem/1162322/inconsistent-behavior-between-llvm-and-std-impleme.html
It's fixed in later versions of LLVM, but not in a way that's easily backported. I haven't noticed any major issues with ignoring the assert, so I'm ignoring it until we can upgrade to newer versions of Clang.