Added debug prints to help investigate issues with floats not getting computed correctly.
Turns out the root cause of the issues I was observing  was the fact that in many places within LLVM uses the address of the fltSemantics instance to figure its identity. There's only a handful of fltSemantics instances and they're all statically allocated. This means that when there's multiple DLLs there's a set of fltSemantics instances for each DLL. We have multiple DLLs (the one that comes with ClangSharp and the one that holds PathogenExtensions) and we process data allocated in one with another, whoops. For whatever reason in certain scenarios the fltSemantics from the ClangSharp libclang were appearing in the APValue instances in libclang-pathogen.

This problem will be properly fixed by https://github.com/InfectedLibraries/llvm-project/issues/2
For now I am manually overriding the native library resolver to only resolve everything to libclang-pathogen.dll